feat: 일정 추가 메뉴 on/off 기능 구현 [TRL-224]
